pyproject.toml
[project]
name = "voice-guard-rail"
version = "0.1.0"
description = "Real-time speech emotion recognition using AI"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.28.0",
    "sounddevice>=0.4.6",
    "soundfile>=0.12.1",
    "numpy>=1.24.0,<2.0.0",
    "torch>=2.6.0",
    "torchaudio>=2.6.0",
    "transformers>=4.35.0",
    "safetensors>=0.4.0",
    "openai-whisper>=20231117",
    "librosa>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]
aws = [
    "boto3>=1.28.0",
]

[project.scripts]
voice-guard-rail = "app:main"

[tool.uv]
dev-dependencies = [
    "pytest>=7.0.0",
    "ruff>=0.1.0",
]
